// src/Rethinktecture/enum/error-type.enum.ts
export enum ErrorEnumType{
    INVALID_ARGUMENT = 1,
    MISSING_ARGUMENT = 2,
    UNAUTHORIZED = 3,
    NOT_FOUND = 4,
    DATABASE_ERROR = 5,
    INTERNAL_ERROR = 6,
    BAD_REQUEST = 400,
    UNAUTHORIZED_HTTP = 401,
    FORBIDDEN = 403,
    NOT_FOUND_HTTP = 404,
    METHOD_NOT_ALLOWED = 405,
    CONFLICT = 409,
    INTERNAL_SERVER_ERROR = 500,
    ERR_NAME_NOT_RESOLVED = 'ERR_NAME_NOT_RESOLVED',
    ERR_INTERNET_DISCONNECTED='ERR_INTERNET_DISCONNECTED',
    ERR_NETWORK='ERR_NETWORK'
}